refactor(ingestion): keep Matchers to matchers, and HTTP out of the classifier (#30149)
Two things this branch got wrong.

Matchers.text was not a matcher. Every other member is a factory returning a
Matcher; text is a projection returning a str, so `when(Matchers.text)` reads as
valid and would match anything truthy. It is now `chain_text`, module-level
beside `exception_chain` - the chain it walks - which also stops it shadowing
`contains(text: str)`.

http_status and response_status were moved into classifier on the claim that
every matcher factory lives there. That claim was wrong: aws.py has held
`aws_code` since before this branch. The rule the package actually follows is
that classifier owns the protocol-agnostic machinery - contains, errno,
exception, any_of walk a cause chain and know nothing else - while a factory
that understands a protocol lives with it. HTTP is a protocol, so `http_status`
and `response_status` sit in checks/rest.py, mirroring aws_code. classifier now
mentions HTTP nowhere.

Their tests move with them.
