+++ ingestion/src/metadata/ingestion/source/dashboard/powerbi/connection.py
@@ -28,9 +28,10 @@ from metadata.core.connections.test_connection import (
 from metadata.core.connections.test_connection.checks.dashboard import DashboardStep
 from metadata.core.connections.test_connection.checks.rest import (
     fetch_list,
+    http_status,
     verify_access,
 )
-from metadata.core.connections.test_connection.classifier import http_status
+from metadata.core.connections.test_connection.classifier import chain_text
 from metadata.core.connections.test_connection.network import NETWORK_ERRORS
 from metadata.generated.schema.entity.services.connections.dashboard.powerBIConnection import (
     PowerBIConnection as PowerBIConnectionConfig,
@@ -58,7 +59,7 @@ def _contains_any(*tokens: str) -> Matcher:
     lowered = tuple(token.lower() for token in tokens)
 
     def match(error: BaseException) -> bool:
-        chain = Matchers.text(error)
+        chain = chain_text(error)
         return any(token in chain for token in lowered)
 
     return match
