+++ ingestion/tests/unit/core/connections/test_connection/test_rest_checks.py
@@ -18,6 +18,8 @@ from metadata.core.connections.test_connection.checks.rest import (
     DEFAULT_LIST_CAP,
     call_endpoint,
     fetch_list,
+    http_status,
+    response_status,
     verify_access,
 )
 from metadata.core.connections.test_connection.records import Diagnosis
@@ -86,3 +88,42 @@ def test_fetch_list_carries_the_command_on_failure():
     with pytest.raises(CheckError) as failure:
         fetch_list(_raise(requests.ConnectionError("down")), noun="job", command="GET /jobs")
     assert failure.value.evidence.command == "GET /jobs"
+
+
+def test_response_status_reads_a_top_level_status_code():
+    error = Exception("api error")
+    error.status_code = 401
+    assert response_status(error) == 401
+
+
+def test_response_status_falls_back_to_the_response():
+    response = requests.Response()
+    response.status_code = 403
+    assert response_status(requests.HTTPError("403", response=response)) == 403
+
+
+def test_response_status_prefers_the_top_level_code_over_the_response():
+    response = requests.Response()
+    response.status_code = 500
+    error = requests.HTTPError("boom", response=response)
+    error.status_code = 401
+    assert response_status(error) == 401
+
+
+def test_response_status_is_none_when_the_error_carries_no_status():
+    assert response_status(Exception("plain")) is None
+
+
+def test_http_status_matches_a_top_level_status_code_across_the_chain():
+    error = Exception("api error")
+    error.status_code = 404
+    wrapper = RuntimeError("wrapped")
+    wrapper.__cause__ = error
+    assert http_status(404)(wrapper) is True
+    assert http_status(401)(wrapper) is False
+
+
+def test_http_status_uses_a_connector_supplied_extractor():
+    error = Exception("vendor 401002")
+    matcher = http_status(401, extract=lambda e: int(str(e).split()[-1][:3]) if "vendor" in str(e) else None)
+    assert matcher(error) is True
